create base directory
