+++ android_tests/lib/android/specs/driver.rb
@@ -34,14 +34,14 @@ describe 'driver' do
       apk_name.must_equal 'api.apk'
     end
 
-    t 'verify Appium::Driver::Capabilities.create_capabilities' do
+    t 'verify Appium::Core::Base::Capabilities.create_capabilities' do
       expected_app = File.absolute_path('../test_apps/api.apk')
-      caps = ::Appium::Core::Driver::Capabilities.create_capabilities(platformName: 'Android',
-                                                                      app:          expected_app,
-                                                                      appPackage:   'io.appium.android.apis',
-                                                                      appActivity:  '.ApiDemos',
-                                                                      deviceName:   'Nexus 7',
-                                                                      some_capability: 'some_capability')
+      caps = ::Appium::Core::Base::Capabilities.create_capabilities(platformName: 'Android',
+                                                                    app:          expected_app,
+                                                                    appPackage:   'io.appium.android.apis',
+                                                                    appActivity:  '.ApiDemos',
+                                                                    deviceName:   'Nexus 7',
+                                                                    some_capability: 'some_capability')
       caps_with_json = JSON.parse(caps.to_json)
       caps_with_json['platformName'].must_equal 'Android'
       caps_with_json['app'].must_equal expected_app
