+++ ios_tests/lib/ios/specs/driver.rb
@@ -30,14 +30,14 @@ describe 'driver' do
     assert_equal expected, actual
   end
 
-  t 'verify Appium::Driver::Capabilities.create_capabilities' do
+  t 'verify Appium::Core::Base::Capabilities.create_capabilities' do
     expected_app = File.absolute_path('UICatalog.app')
-    caps = ::Appium::Core::Driver::Capabilities.create_capabilities(platformName:    'ios',
-                                                                    platformVersion: '10.3',
-                                                                    automationName:  'XCUITest',
-                                                                    deviceName:      'iPhone Simulator',
-                                                                    app:             expected_app,
-                                                                    some_capability: 'some_capability')
+    caps = ::Appium::Core::Base::Capabilities.create_capabilities(platformName:    'ios',
+                                                                  platformVersion: '10.3',
+                                                                  automationName:  'XCUITest',
+                                                                  deviceName:      'iPhone Simulator',
+                                                                  app:             expected_app,
+                                                                  some_capability: 'some_capability')
     caps_with_json = JSON.parse(caps.to_json)
     caps_with_json['platformName'].must_equal 'ios'
     caps_with_json['platformVersion'].must_equal '10.3'
