+++ lib/appium_lib/core/common/base.rb
@@ -1,0 +1,3 @@
+require_relative 'base/base'
+require_relative 'base/bridge'
+require_relative 'base/capabilities'
