+++ lib/appium_lib/core/common/base/bridge.rb
@@ -1,32 +1,6 @@
-require_relative 'common/search_context'
-
-# TODO: should clean
-
 module Appium
   module Core
     class Base
-      class Driver < ::Selenium::WebDriver::Driver
-        include ::Selenium::WebDriver::DriverExtensions::UploadsFiles
-        include ::Selenium::WebDriver::DriverExtensions::TakesScreenshot
-        include ::Selenium::WebDriver::DriverExtensions::HasSessionId
-        include ::Selenium::WebDriver::DriverExtensions::Rotatable
-        include ::Selenium::WebDriver::DriverExtensions::HasRemoteStatus
-        include ::Selenium::WebDriver::DriverExtensions::HasWebStorage
-
-        include ::Appium::Core::SearchContext
-
-        def initialize(opts = {})
-          listener = opts.delete(:listener)
-          @bridge = ::Appium::Core::Base::Bridge.handshake(opts)
-          if @bridge.dialect == :oss
-            extend ::Selenium::WebDriver::DriverExtensions::HasTouchScreen
-            extend ::Selenium::WebDriver::DriverExtensions::HasLocation
-            extend ::Selenium::WebDriver::DriverExtensions::HasNetworkConnection
-          end
-          super(@bridge, listener: listener)
-        end
-      end # class Driver
-
       class Bridge < ::Selenium::WebDriver::Remote::Bridge
         def self.handshake(**opts)
           desired_capabilities = opts.delete(:desired_capabilities)
