+++ lib/appium_lib/core/common/device.rb
